+++ Generals/Code/Tools/WorldBuilder/include/GlobalLightOptions.h
@@ -92,7 +92,7 @@ private:
 			if (component <= 0) {
 				return 0.0;
 			}
-			percent = (Real)component/255.0;
+			percent = (Real)component/255.0f;
 			return percent;
 	}
 	Int		PercentToComponent(Real percent)
@@ -104,7 +104,7 @@ private:
 			if (percent <= 0.0) {
 				return 0;
 			}
-			component = (percent * 255.0);
+			component = (Int)(percent * 255.0f);
 			return component;
 	};
 	BOOL	GetInt(Int ctrlID, Int *rVal);
